[add] entropy
